requirements.txt
neat-python
pygame
random